+++ packages/server/src/app/controllers/content/index.ts
@@ -4,6 +4,7 @@ import { MessagesController } from './messages.controller';
 import { CommentsController } from './comments.controller';
 import { CollectionsController } from './collections.controller';
 import { SectionsController } from './sections.controller';
+import { PortfolioController } from './portfolio.controller';
 
 export const ContentRoutes = [
     ContentController,
@@ -12,4 +13,5 @@ export const ContentRoutes = [
     CommentsController,
     CollectionsController,
     SectionsController,
+    PortfolioController,
 ];
