--- packages/server/src/app/controllers/portfolio/index.ts
@@ -1,3 +1,0 @@
-import { PortfolioController } from './portfolio.controller';
-
-export const PortfolioRoutes = [PortfolioController];
