--- packages/server/src/app/controllers/portfolio/portfolio.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('portfolio')
-export class PortfolioController {}
